Adaptation to new standards
git-svn-id: file:///home/roberto/Git/svn@5617 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 1508610fba97c64a4ce09e65e1fe6618d2dceeb2

+++ trunk/xmipp/libraries/reconstruction/sort_images.cpp
@@ -169,7 +169,7 @@ void Prog_sort_images_prm::run()
             SFInfo.setValue(MDL_IMAGE,fnOut);
             SFInfo.setValue(MDL_IMAGE_ORIGINAL,fnOutOrig);
             SFInfo.setValue(MDL_IMAGE_CLASS_GROUP,fnSel);
-            SFInfo.setValue(MDL_IMAGE_CLASS_COUNT, SFaux.size());
+            SFInfo.setValue(MDL_IMAGE_CLASS_COUNT, (int) SFaux.size());
             SFoutOriginal.nextObject();
         }
         SFInfo.write(fnRoot+"_info.txt");
